+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.1.1
 * Define `UT` joint scheme
+* Define joint scheme for `Maybe` datatype